==> _script/d-experiment/13_create_concept.ipynb <==
# create concept framework using raw data
## use Boston

# load boston hexagon and create the segmentation summary
import geopandas as gpd
import pandas as pd
import numpy as np
import os
import sys
from glob import glob
ROOT_DIR = "../"
sys.path.append(ROOT_DIR)
import matplotlib.pyplot as plt
import seaborn as sns
import h3
# city_meta = load_all()
plt.rcParams["figure.figsize"] = [3, 4]
plt.rcParams["figure.autolayout"] = True
plt.rcParams['svg.fonttype'] = 'none'


# load geojson
folder = r"D:\Dropbox (Personal)\Personal Work\_Projects2023\01_city-never-was\_data\_curated\c_hex_cluster\cluster=8"
folder_2 = r"D:\Dropbox (Personal)\Personal Work\_Projects2023\01_city-never-was\_data\_curated\c_analysis\seg_summary"
boston_hex = gpd.read_file(os.path.join(folder, "boston_cluster=8-tsn-res=9.geojson"))
boston_seg = pd.read_csv(os.path.join(folder_2, "boston_seg_summary.csv"))
boston_hex.head()

boston_hex_viz = boston_hex.merge(boston_seg, 
                                  left_on="hex_id",
                                    right_on="h3_9",
                                  )
boston_hex_viz.head()

boston_hex_viz.columns

exportfolder = r"D:\Dropbox (Personal)\Personal Work\_Projects2023\01_city-never-was\_data\_curated\c_analysis\for_conceptual"
boston_hex_viz[['cluster_8', 'geometry', 'h3_9', 
        'img', 'panoid', 'sidewalk_presence',
       'exposure_presence','obj_motorcycle', 'obj_person']]\
           .to_file(os.path.join(exportfolder, "boston_hex_seg_summary.geojson"), driver='GeoJSON')
